CodeWars/Java/src/test/java/ReverseWordsTests.java
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class ReverseWordsTests {
    @Test
    @DisplayName("Should return a string with all its words reversed")
    public void testReverseWords() {
        assertEquals("", ReverseWords.reverseWords(""));
        assertEquals("    ", ReverseWords.reverseWords("    "));
        assertEquals("ehT kciuq nworb xof spmuj revo eht yzal .god", ReverseWords.reverseWords("The quick brown fox jumps over the lazy dog."));
        assertEquals("elppa", ReverseWords.reverseWords("apple"));
        assertEquals("a b c d", ReverseWords.reverseWords("a b c d"));
        assertEquals("elbuod  decaps  sdrow", ReverseWords.reverseWords("double  spaced  words"));
    }
}
